projects/04-rag-from-scratch/requirements.txt
sentence-transformers==3.0.1
chromadb==0.5.3
rich==13.7.1
requests==2.32.3
pyyaml==6.0.2
boto3==1.35.0
pytest==8.3.2
